+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWKHQSRTFP8MCJC3Q967QVQH"
+        "01KWKPBP785VWY9RZ6HHW9KJG4"
     )
     r"""The unique identifier of the data source"""