+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-03T08:32:24.507Z"
+    created: Optional[str] = "2026-07-03T09:53:14.266Z"
 
-    updated: Optional[str] = "2026-07-03T08:32:24.507Z"
+    updated: Optional[str] = "2026-07-03T09:53:14.266Z"
 
     updated_by_id: OptionalNullable[str] = UNSET