+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWKHQSRXN29NBNEDS35PSYJS"
+        "01KWKPBP7DMHDZC6CTRCJ5H9SN"
     )
     r"""The unique identifier of the data source"""